+++ src/main/java/ar/edu/unlp/info/oo1/practica1/ejercicio2/Interfaces/ITicket.java
@@ -4,11 +4,11 @@ import java.time.LocalDate;
 
 public interface ITicket
 {
-    float impuesto();
+    double impuesto();
 
-    float getPrecioTotal();
+    double getPrecioTotal();
 
-    float getPesoTotal();
+    double getPesoTotal();
 
     int getCantidadDeProductos();
